+++ R/aggregate.R
@@ -84,7 +84,7 @@ aggregate.stars = function(x, by, FUN, ..., drop = FALSE, join = st_intersects,
 		geom = "geometry"
 	stopifnot(!missing(FUN), is.function(FUN))
 
-	if (exact && inherits(by, c("sf", "sfc_POLYGON", "sfc_MULTIPOLYGON")) && has_raster(x)) {
+	if (exact && inherits(by, c("sfc_POLYGON", "sfc_MULTIPOLYGON")) && has_raster(x)) {
     	if (!requireNamespace("raster", quietly = TRUE))
         	stop("package raster required, please install it first") # nocov
     	if (!requireNamespace("exactextractr", quietly = TRUE))
@@ -95,8 +95,7 @@ aggregate.stars = function(x, by, FUN, ..., drop = FALSE, join = st_intersects,
 		e = exactextractr::coverage_fraction(as(r, "Raster"), by)
 		st = do.call(raster::stack, e)
 		m = raster::getValues(st)
-		#if (identical(FUN, mean)) see https://github.com/r-spatial/stars/issues/289
-		if (!identical(FUN, sum)) {
+		if (!identical(FUN, sum)) { # see https://github.com/r-spatial/stars/issues/289
 			if (isTRUE(as.character(as.list(FUN)[[3]])[2] == "mean"))
 				m = sweep(m, 2, colSums(m), "/") # mean: divide weights by the sum of weights
 			else