+++ NEWS.md
@@ -1,5 +1,7 @@
 # version 0.6-2
 
+* `[<-.stars()` accepts for `i` an indicator (numeric length character) to the array to be replaced; #626
+
 * `plot.stars()` gains an argument `key.lab` to set the legend key label (requires sf >= 1.0-13)
 
 # version 0.6-1